week_2/day_4/exercises_xp/Restaurant manger/pyproject.toml
[project]
name = "restaurant-manger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
